Move ui to ui/buffer.lua
